ensuring correct value output with casting

+++ coms/testprogs/joystickserialreceiver/joystickserialreceiver.cpp
@@ -16,8 +16,8 @@ int main()
 	{
 	  if(js_receiver.readSerialEvent(js_event))
 	    {
-	      std::cout << "time [ms]: " << js_event.time_ms << std::endl;
-	      std::cout << "value: " << js_event.value << std::endl;
+	      std::cout << "time [ms]: " << (int)js_event.time_ms << std::endl;
+	      std::cout << "value: " << (int)js_event.value << std::endl;
 	      std::cout << "type: " << (int)js_event.type << std::endl;
 	      std::cout << "number: " << (int)js_event.number << std::endl;
 	    }
